refactor: split modules into dedicated files
